* added end date for recurrences ("<range>" to "UNTIL")
This currently works for ranges with type=date. Other range types are
not supported (yet). The end date in Kolab is valid for the whole day.
In RFC 2445 UNTIL specifies the last valid second for a recurring event.
So the date is converted to "date 23:59:59".

* added exceptions for recurrences ("<exclusion>" to "exDate")

Via exDate (RFC2445) a list of exclusion dates is specified generated
by converting "<exclusion>". If the recurring event is not a whole day
event the exclusion date has to contain the exact date and time of the
recurrence (thus copied here if not "start.allDay").

git-svn-id: https://kolab-android.googlecode.com/svn/trunk@73 16dfca22-1d42-11df-8821-87deac751e95